Send/Recieve arbitrary signals
This adds the ability to both send and recieve arbitrary signals on the bus.
To recieve an arbitrary signal, you must implement DBusSigHandler(as normal),
but the generic type is just DBusSignal.  This adds two new methods to help
accomplish this on AbstractConnection: addGenericHandler and
removeGenericHandler.  The new methods are required in order for the code
that will normally create a new subclass of DBusSignal, which for arbitrary
signals we can't have.