Removed timeout on inputStream read.
Reading on the InputStream should always wait for an infinite time.
Stop reading after a given (fixed) timeout will cause strange
application behaviour and also is not the way Dbus is working.

On DBus there are no heartbeats, so it is completely valid that there is
no message for a long time - until the registered application wants to
tell something to the registered consumers.

The only timeout which makes sense is a connection timeout when using
TCP instead of UnixSocket (FileDescriptor).

On UnixSockets there is no connect timeout. The connection will work
or will fail instantly (due to wrong permissions or socket
file not found).