Build dbus-java as bundle with embedded dependencies.
The challenge with building this project as an OSGi-enabled bundle is that many of the dependencies it
relies on are also not OSGi-enabled.  There are a couple ways to manage this, but the two broad categories
are to either make those dependency projects OSGi-enabled, or to take those projects as-is and embed
them in the bundle that is created.  This PR is one of two different proposals to build dbus-java as a bundle.
This version of pom.xml will embed all the dependencies needed for this bundle to be loaded and run,
including jffi-*.jar, jnr-*.jar, and asm-*.jar.  The benefit to this method is convenience for users
who want to reference this library an an OSGi environment-it's ready to
go.  The drawback is the bundle becomes "fat" by embedding other jar files in the bundle itself.