Fixed regression in finding correct Constructor (#114).
The changes introduced in #110 did not cover the use case of interfaces classes
used as constructor arguments. When e.g. using Map or List interface in constructor the
current expectedArgumentsList.equals(receivedArgumentsList) will fail, because the received arguments
are not the interface type, but a compatible class.
This requires some more attention, which means every expectedArgument has be be changed for
compatibility to the received argument type.

This behavior has been added in this commit.
Every received argument is now checked with 'isAssignable' to the expected arguments.
This will allow inheritance for objects as well as allowing the usage of interface classes in constructor
definitions.