Modularise dbus-java (MRJAR, so supports Java 8 and Java 9+).
Now that JNR at least uses Automatic-Module-Name, dbus-java itself can modularised.  The largest change was the need to rename org.freedesktop.dbus.bin to org.freedesktop.dbus.utils.bin. No more than one particular module may export the same package, and both dbus-java-utils and dbus-java tried to export org.freedesktop.dbus.bin.

I should also point our that having classes in the org.freedesktop namespace is probably a bad idea (org.freedesktop.DBus and org.freedesktop.Hexdump classes), as this would mean that no modular project could use dbus-java and another module that happens to export org.freedesktop. One of my own projects did this (jfreedesktop), and exposed the issue. It can be left as org.freedesktop for now, but to be a good modular citizen this should be changed.

The other modular imperfect is `java.utils`. This is using an automatic module name, and should itself be modularised.

The Jar produced by the build is a multi-release Jar. So the bulk of it is compiled to Java 8, but Java 9 code is also compiled and placed in META-INF/versions/9. This is of course only used for module-info.class.