dbus-java-utils: ClassBuilderInfo: Don't duplicate constructor arguments
When building classes using ClassBuilderInfo, arguments to the super()
[superArguments] are automatically added to the constructor of the newly
generated class. These arguments are not saved to the private members of
the class. Saving to private members in the constructor is only done for
[arguments] only.
To have both:
* saving to private members for getters and setters and
* supplying to super()
we have to add the argument to both: superArguments and arguments. But
this then results to duplicate arguments to the constructor.
This commit now removes these double arguments.