dbus-java-utils: InterfaceCodeGenerator: Fix constructor for signals
The constructor for dbus signals was generated wrong. For example this
introspection.xml:

<node>
  <interface name="de.Example">
    <signal name="ExampleSignal">
      <arg type="s" name="newState"/>
    </signal>
  </interface>
</node>

did generate this class:

public interface Example extends DBusInterface {
    public static class ExampleSignal extends DBusSignal {

        private final String newState;

        public ExampleSignal(String _path, String _interfaceName, String _newState) throws DBusException {
            super(_path, _interfaceName);
            this.newState = _newState;
        }

        public String getNewState() {
            return newState;
        }
    }
}

This is wrong. With this interface it is not possible to emit this
signal using dbus-java.
When the ExampleSignal constructor is called, this leads to the
DBusSignal constructor to be called with _path and _interfaceName
arguments. DBusSignal class then infers from the enclosing classes
arguments, that there must be two string arguments ("ss") to this
signal. It is only one.

The interface constructor has to look something like this:

        public ExampleSignal(String _path, String _newState) throws DBusException {
            super(_path, _newState);
            this.newState = _newState;
        }

This is what this commit does. It changes the interface generation to
reflect this.