InterfaceCodeGenerator: Fix methods without direction on args
DBus specification does allow methods that do not have the "direction"
attribute on their args. These are threatened as "in".
The InferfaceCodeGenerator did ignore method args that do not have a
direction set and no arguments where created on the java interface.
With this commit now method args without a direction attribute are also
threatened as if they'd had direction="in" set. Other values than "in"
or "out" are still ignored.