InterfaceCodeGenerator: Fix *Tuple return type for methods
Generated methods that have multiple "out" args solve this by having a
*Tuple container class in java.
This did not work. The generated return type of the method was the type
of the first "out" arg in dbus.
This is now fixed now the *Tuple class is the return type of a methode
with multiple return values.