Add possibility for Tuples as return types to dbus method calls
In dbus it is possible for a method to have multiple return types. To
map this into the java world, the dbus-java library wraps the return
types into a class, that contains all desired values as fields. This
class extends the Tuple class and is normally generated by the
InterfaceCodeGenerator.
There was a piece missing in the library, that could deal with such
Tuples as a return type to a dbus method call.
This commit adds the missing piece and a little test case.