Fixed issue introduced with PR #143:
Using HashSet and putting methods to a map without checking if this is a
duplicate will cause random failures in TestAll test.

TestAll has a test case which will override the 'Introspect' method with
some bogus method.

ExportedObject will create the Introspection data and will maintain a
list of methods which can be called on the containing object.
It also adds Peer and Introspectable to every exported object.

When a method of Introspectable (or Peer) is overridden by another
interface, HashSet will cause random failures because the export order
is not retained. Using LinkedHashSet here will fix the order issue,
because the methods of the given class/interface is added first, and the
additional methods pulled in from Introspection and Peer interface is
added last.

It is also required to check if a method is already part of the
'methods' map, otherwise the 'good' method will be overridden by the
'bad' method because without check the later always wins.