Improved signal handling
This change will remove the usage of SignalTuple class as map key in
AbstractConnection.
It will also remove the need to create 12+ SignalTuple object for every
handled signal to maybe find a proper signal handler in internal handler
map.
Instead all signal handlers are stored using the (already
created/existing) DBusMatchRule.
DBusMatchRule was improved to contain hashcode and equals methods to
work as map key properly.

Also two methods were added to DBusMatchRule (matches(DBusMatchRule,
boolean) and matches(DBusSignal, boolean)).
Those methods will be used to match a rule against a signal or other
rule using strict or partial matching.
Partial means that only the parts of the rule are matched to the signal
which were non-null in the rule.

When a signal is received the internal map will be iterated looking for
a partial matching rule. If any, all handlers assigned to that rule will
be called. A signal can match multiple rules (depending on how 'wide'
the rule was defined), in any case all matching handlers will be called.