BUGFIX: add the interface type if given
The interface type value "_type" is already null if not given as a parameter. But if given, it's being ignored while constructing the RemoteObject reference that will eventually be proxied. This leads to method invocations failing with "strange" errors (access denied).

This correction caused the methods to at least be invoked correctly.