Modified the POM files to be more CI/CD friendly
In the new CI/CD world, it's easier to set a version for a specific build by just setting -Drevision=... and invoking the build. This significantly facilitates the whole versioning paradigm since the version is now only stored in a single, centralized place.

Now, whenever a version bump is needed, only the "master" value in the root pom.xml needs to be updated. This should only affect the build process.

See https://maven.apache.org/maven-ci-friendly.html for more details on ${revision}, ${changelist}, and ${sha1} and their usefulness in the modern, CI/CD world.