Improvements of Message class
This change includes several changes at once.

1) Use Object[] instead of a Map for the message header

The header was stored in a Map<Byte, Object>, where the byte key
defines the header field number in the actual message.

Using Map as container provided a nicer API but creates additional
overhead. Also the get() calls on a Map always uses hashCode() to find
the proper key.

Additionally, changes made to the header Map were never written back
to the message.

As the key is just a byte and the header of a DBus message only has 10
entries, an array of Object is the better option.
The array index defines the position in the message header. Index 0 is
omitted (invalid according to DBus Spec).
Looking up values using the array index is a lot faster.

2) Fix issues when running as daemon (Issue #207)

When the Java DBusDaemon was used and a second program exported methods to this
DBusDaemon, requesting results from these methods using dbus-send
failed.

This was due to the message header was corrupted when a message was
forwarded to the requester.

Updating the target for a message was done using setSource() method
which was (and still is) a very nasty beast.
First it clears the current buffers, then rewrites the initial part of
the message. After that, the header is rewritten (that part was wrong
since the beginning). After the header the body bytes are appendend.

The broken part did remove the 'signature' indicating byte for any
MethodReturn call.