Improvements in DBusDaemon
This commit changes several problems in DBusDaemon.
First, the configured transport was not used properly.
In case a different MessageReader or Writer was present on classpath,
the server used the default MessageReader/Writer.

This was due to the fact that a listening connection will create multiple SocketChannel (one
for each client) and the created MessageReader/Writer was set in the
Transport.
The transport object was never used inside of the DBusDaemon, just the
SocketChannel was passed to the daemon.

This required some changes in AbstractTransport, and will also
change the behavior when trying to "connect" a listening address.

In AbstractTransport a new meta object 'TransportConnection' is created.
It bundles the SocketChannel, MessageReader and MessageWriter.
This object is created and returned when createInputOutput() method is called internally.

When using the 'connect' call, everything behaves like before except
when the bus addresses 'isListening()' method returns true, a
InvalidBusAddressException is thrown.
This is because creating MessageReader/Writer for any
connecting client is pointless as it would never been used.

To use a listening address 'listen()' has to be used.
This will in turn throw InvalidBusAddressException when the address
is not a listening address. It will also return the TransportConnection
so the caller can use the configured reader and writer properly.

DBusDaemon was broken because it did not wait for the read/send thread
to be properly initialized before continuing. This was caused by wrong
run-loop conditions (or better the condition was set at the wrong
place). Also the listening-loop should be a do...while loop so the
listening part is at least called once.

In addition to that, the "localqueue" could be removed and DBusServer
does not need to be a subclass of Thread.