Support connection based endianess
This change is quite large.

Every message needs to know which endianess it should contain.
As soon as the message content is serialized to the approriate byte
array, the endianess can no longer be changed.

Therefore, all messages should be created using the MessageFactory
provided by the AbstractConnection instance.
The factory already exists in previous versions but was only used when
creating signals and method return calls.
Now it has been extended to create all sorts of messages.
This was done to prevent changes to every Message-based subclass
constructor.

For testing purposes it is still possible to use
AbstractConnection.sendMessage(Message) to actually send any Message
compatible message.
It is required that each message passed to this method has a proper
endianess set manually if the message was not created using the message factory.

As DBusSignals are usually implemented in user code, it may be possible
that these signals use public constructors and therefore will never set
the endianess. To fix this, signals send by sendMessage(Message) call
will update the endianess before updating.
This can be done because Messages created by the constructor were not
yet serialized.

This change also moves the Error class to the messages
package because Error is a subclass of Message and only the message
factory should be able to create new instances of Error.